backend/requirements.txt
# 核心框架
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# 数据库相关
sqlalchemy==2.0.23
pymysql==1.1.0

# 数据处理
pandas==2.1.4
numpy==1.25.2

# 机器学习
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.3.0
scipy==1.11.4

# 数据可视化
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.2

# 中文分词和文本处理
jieba==0.42.1
textblob==0.17.1
snownlp==0.12.3

# 网络请求
requests==2.31.0

# 进度条
tqdm==4.66.1

# 任务调度
apscheduler==3.10.4

# 图像处理
pillow==10.1.0

# 认证相关
passlib==1.7.4
bcrypt==4.1.2
pyjwt==2.8.0
python-multipart==0.0.6

# CORS支持
python-cors==1.7.0

# AI服务
openai==1.58.1
jinja2==3.1.2
matplotlib==3.8.2
seaborn==0.13.0 